+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 - fix(bitcoin-da): Harden DA monitoring reveal restore ([#3329](https://github.com/chainwayxyz/citrea/pull/3329))
 - fix(rpc): Set archival state in estimate tx expenses ([#3321](https://github.com/chainwayxyz/citrea/pull/3321))
+- fix(sequencer): Keep producing L2 blocks on stale DA tip ([#3324](https://github.com/chainwayxyz/citrea/pull/3324))
 - fix(rpc): with log config in debug trace cache ([#3323](https://github.com/chainwayxyz/citrea/pull/3323))
 
 ## [v2.7.0](2026-07-29)